requirements.txt
# Core dependencies
torch==2.2.1+cu121
torchvision==0.17.1+cu121
torchaudio==2.2.1+cu121
onnxruntime-gpu==1.21.1
ultralytics==8.1.28
opencv-python==4.9.0.80
numpy==1.26.4
flask==3.0.2

# CUDA dependencies
cuda-python==12.1.0
cudnn==8.9.2.26

# Additional dependencies
pillow==10.2.0
requests==2.31.0
tqdm==4.66.2
matplotlib==3.8.3
pandas==2.2.1
scikit-learn==1.4.1

# Web interface dependencies
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0

# Development dependencies
pytest==8.0.2
black==24.2.0
flake8==7.0.0

# Build dependencies
cmake>=3.27.7
protobuf>=4.25.1

# Deep Learning Framework Dependencies
--extra-index-url https://download.pytorch.org/whl/cu118
tensorflow==2.15.0
onnx==1.14.1

# Image Processing Libraries
scipy>=1.11.4
scikit-image>=0.22.0
albumentations>=1.3.1

# Web Application Dependencies
werkzeug>=3.0.1

# System Utilities
psutil>=5.9.7
py-cpuinfo>=9.0.0 